Revert "broke router into modules, reorganized server.js, created index.js"
