--- index.js
@@ -1,4 +1,0 @@
-'use strict';
-
-require('dotenv').config();
-require('./lib/server.js').start();
